Added new vanilla-bootstrap webpage allowing users to modify their status
The webpage allows users to change their status given that they know the password. The webpages are respectiviely hosted by the individual pi's. When a modification is made, (e.g. a status has changed) the updated information will be sent to the master pi's database. Files have been cleaned up to use snake case and they are all located in one directory for simplicity.